# apps/feature-flags-api/docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Read-only boolean feature flags served from repo JSON (local/prod).",
        "title": "Feature Flags API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
        "/admin/reload": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Reload flags from disk (internal use)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/flags": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get all flags for an environment",
                "parameters": [
                    {
                        "enum": [
                            "local",
                            "prod"
                        ],
                        "type": "string",
                        "description": "Environment",
                        "name": "env",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/flags/{key}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get a single flag’s status for an environment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Flag key (snake_case)",
                        "name": "key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "local",
                            "prod"
                        ],
                        "type": "string",
                        "description": "Environment",
                        "name": "env",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.FlagStatus"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "handlers.FlagStatus": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                }
            }
        }
    }
}